requirements.txt
fastapi
uvicorn[standard]
pillow
numpy==1.26.4
python-multipart
pydantic
git+https://github.com/openai/CLIP.git
